chore: major project restructure
Data is now inside lib, being available on page generation
Multiple components abstracted into components folder
